Decouple the pool implementation from the memory subsystem.
