+++ src/CuArrays.jl
@@ -112,7 +112,6 @@ function __init__()
     end
 
     __init_memory__()
-    __init_pool_()
 end
 
 end # module
